+++ tellMe/src/main/java/com/bigProject/tellMe/repository/NoticeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface NoticeRepository extends CrudRepository<Notice, Integer> {
+public interface NoticeRepository extends CrudRepository<Notice, Long> {
 }
